requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0

# Chess libraries
python-chess>=1.999
python-lichess>=0.10

# OpenAI API
openai>=1.0.0

# HTTP requests
requests>=2.28.0

# Environment variables
python-dotenv>=0.19.0

# Optional: for data visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Optional: for Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.0.0

# Development dependencies (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
